scripts/draft/requirements.txt
anthropic>=0.40.0
PyYAML>=6.0
jsonschema>=4.0
